+++ packages/gatsby-dev-cli/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.0-beta.4"></a>
+
+# [2.0.0-beta.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-dev-cli/compare/gatsby-dev-cli@2.0.0-beta.3...gatsby-dev-cli@2.0.0-beta.4) (2018-07-27)
+
+**Note:** Version bump only for package gatsby-dev-cli
+
+<a name="2.0.0-beta.3"></a>
+
+# [2.0.0-beta.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-dev-cli/compare/gatsby-dev-cli@2.0.0-beta.2...gatsby-dev-cli@2.0.0-beta.3) (2018-07-21)
+
+**Note:** Version bump only for package gatsby-dev-cli
+
 <a name="2.0.0-beta.2"></a>
 
 # [2.0.0-beta.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-dev-cli/compare/gatsby-dev-cli@2.0.0-beta.1...gatsby-dev-cli@2.0.0-beta.2) (2018-06-20)