+++ packages/gatsby-plugin-react-css-modules/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.0-beta.4"></a>
+
+# [2.0.0-beta.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-css-modules/compare/gatsby-plugin-react-css-modules@2.0.0-beta.3...gatsby-plugin-react-css-modules@2.0.0-beta.4) (2018-07-21)
+
+**Note:** Version bump only for package gatsby-plugin-react-css-modules
+
 <a name="2.0.0-beta.3"></a>
 
 # [2.0.0-beta.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-css-modules/compare/gatsby-plugin-react-css-modules@2.0.0-beta.2...gatsby-plugin-react-css-modules@2.0.0-beta.3) (2018-06-21)